Rabi fitting for data from glue
Read in ekkorrection.dat and lpbfit.dat. Based on the results from lpbfit, search for the minimum absorption value in the vicinity. Using this minimum value, fit the upper and lower energy branches of the strongly coupled Rabi splitting. Note that the file specifies the photon linewidth and exciton linewidth as 38.1 meV and 51.8 meV respectively. Export the extracted minimum value and the fitted upper and lower energy branch results.